requirements.txt
streamlit
google-generativeai>=0.7.2
PyMuPDF
reportlab
python-dotenv
